Fix (remove) dark theme.  Move components to their own files.  Make albums bookmarkable.
